+++ src/main/java/com/andersmmg/lockandblock/LockAndBlock.java
@@ -2,6 +2,7 @@ package com.andersmmg.lockandblock;
 
 import com.andersmmg.lockandblock.block.ModBlocks;
 import com.andersmmg.lockandblock.block.entity.ModBlockEntities;
+import com.andersmmg.lockandblock.config.ModConfig;
 import com.andersmmg.lockandblock.item.ModItemGroups;
 import com.andersmmg.lockandblock.item.ModItems;
 import com.andersmmg.lockandblock.sounds.ModSounds;
@@ -20,7 +21,7 @@ import org.slf4j.LoggerFactory;
 public class LockAndBlock implements ModInitializer {
     public static final String MOD_ID = "lockandblock";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
-    //    public static final ModConfig CONFIG = ModConfig.createAndLoad();
+    public static final ModConfig CONFIG = ModConfig.createAndLoad();
     public static final String CARD_UUID_KEY = "card_uuid";
 
     public static final RegistryKey<DamageType> TESLA_COIL_DAMAGE = RegistryKey.of(RegistryKeys.DAMAGE_TYPE, id("tesla_coil_damage_type"));