+++ CHANGELOG.md
@@ -8,8 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - Tesla coil
+- Mod config options
 
 ### Changed
+
+- Updated metadata
 
 ### Deprecated